fix(client): jsdoc for undefined vs null. Closes #5149
